+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- new API `JsonValidationService.newInstance(JsonProvider)` 
+
 ## 3.0.0-RC1 - 2020-04-27
 ### Changed
 - Updated Jakarta JSON Processing API to 2.0.0, which causes the API package `javax.json` to be migrated to `jakarta.json`.